fix out of proc factory caching, add test

+++ src/Tasks/RoslynCodeTaskFactory/RoslynCodeTaskFactory.cs
@@ -84,7 +84,19 @@ namespace Microsoft.Build.Tasks
         /// A cache of <see cref="RoslynCodeTaskFactoryTaskInfo"/> objects and their corresponding compiled assembly.  This cache ensures that two of the exact same code task
         /// declarations are not compiled multiple times.
         /// </summary>
-        private static readonly ConcurrentDictionary<RoslynCodeTaskFactoryTaskInfo, Assembly> CompiledAssemblyCache = new ConcurrentDictionary<RoslynCodeTaskFactoryTaskInfo, Assembly>();
+        private static readonly ConcurrentDictionary<RoslynCodeTaskFactoryTaskInfo, CachedAssemblyEntry> CompiledAssemblyCache = new ConcurrentDictionary<RoslynCodeTaskFactoryTaskInfo, CachedAssemblyEntry>();
+        private readonly struct CachedAssemblyEntry
+        {
+            public CachedAssemblyEntry(Assembly assembly, string assemblyPath)
+            {
+                Assembly = assembly;
+                AssemblyPath = assemblyPath;
+            }
+
+            public Assembly Assembly { get; }
+
+            public string AssemblyPath { get; }
+        }
 
         /// <summary>
         /// Stores the path to the directory that this assembly is located in.
@@ -144,7 +156,7 @@ namespace Microsoft.Build.Tasks
             return _parameters;
         }
 
-        /// <inheritdoc cref="IOutOfProcTaskFactory.GetAssemblyPath"/>
+        /// <inheritdoc cref="IOutOfProcTaskFactory"/>
         public string GetAssemblyPath() => _assemblyPath;
 
         /// <inheritdoc cref="ITaskFactory.Initialize"/>
@@ -663,28 +675,47 @@ namespace Microsoft.Build.Tasks
         /// <returns><code>true</code> if the source code could be compiled and loaded, otherwise <code>false</code>.</returns>
         private bool TryCompileAssembly(IBuildEngine buildEngine, RoslynCodeTaskFactoryTaskInfo taskInfo, out Assembly assembly)
         {
-            // First attempt to get a compiled assembly from the cache
-            if (CompiledAssemblyCache.TryGetValue(taskInfo, out assembly))
-            {
-                return true;
-            }
+            CachedAssemblyEntry cachedEntry;
+            bool reuseCompiledAssembly = CompiledAssemblyCache.TryGetValue(taskInfo, out cachedEntry);
+            assembly = reuseCompiledAssembly ? cachedEntry.Assembly : null;
+            string? sourceCodePath = null;
 
-            // The source code cannot actually be compiled "in memory" so instead the source code is written to disk in
-            // the temp folder as well as the assembly. After build, the source code and assembly are deleted.
-            string sourceCodePath = FileUtilities.GetTemporaryFileName(".tmp");
-
-            if (Traits.Instance.ForceTaskFactoryOutOfProc)
+            if (reuseCompiledAssembly)
             {
-                _assemblyPath = TaskFactoryUtilities.GetTemporaryTaskAssemblyPath(); // in a temp directory for this process, persisted until the end of build
+                string cachedPath = cachedEntry.AssemblyPath;
+                if (!string.IsNullOrEmpty(cachedPath))
+                {
+                    _assemblyPath = cachedPath;
+                }
+                else
+                {
+                    _assemblyPath = assembly?.Location;
+                }
             }
             else
             {
-                _assemblyPath = FileUtilities.GetTemporaryFileName(".dll"); // dll in the root of the temp directory, removed immediately after compilation
+                // The source code cannot actually be compiled "in memory" so instead the source code is written to disk in
+                // the temp folder as well as the assembly. After build, the source code and assembly are deleted.
+                sourceCodePath = FileUtilities.GetTemporaryFileName(".tmp");
+
+                if (Traits.Instance.ForceTaskFactoryOutOfProc)
+                {
+                    _assemblyPath = TaskFactoryUtilities.GetTemporaryTaskAssemblyPath(); // in a temp directory for this process, persisted until the end of build
+                }
+                else
+                {
+                    _assemblyPath = FileUtilities.GetTemporaryFileName(".dll"); // dll in the root of the temp directory, removed immediately after compilation
+                }
             }
 
             if (!TryResolveAssemblyReferences(_log, taskInfo, out ITaskItem[] references))
             {
                 return false;
+            }
+
+            if (reuseCompiledAssembly)
+            {
+                return true;
             }
 
             // Delete the code file unless compilation failed or the environment variable MSBUILDLOGCODETASKFACTORYOUTPUT
@@ -770,7 +801,8 @@ namespace Microsoft.Build.Tasks
                 // Return the compiled assembly
                 assembly = TaskFactoryUtilities.LoadTaskAssembly(_assemblyPath);
 
-                CompiledAssemblyCache.TryAdd(taskInfo, assembly);
+                string cachedAssemblyPath = Traits.Instance.ForceTaskFactoryOutOfProc ? _assemblyPath : string.Empty;
+                CompiledAssemblyCache.TryAdd(taskInfo, new CachedAssemblyEntry(assembly, cachedAssemblyPath));
                 return true;
             }
             catch (Exception e)
@@ -780,14 +812,15 @@ namespace Microsoft.Build.Tasks
             }
             finally
             {
-                if (deleteSourceCodeFile && FileSystems.Default.FileExists(sourceCodePath))
+                if (deleteSourceCodeFile && sourceCodePath is not null && FileSystems.Default.FileExists(sourceCodePath))
                 {
                     File.Delete(sourceCodePath);
                 }
 
-                if (!Traits.Instance.ForceTaskFactoryOutOfProc && FileSystems.Default.FileExists(_assemblyPath))
+                if (!Traits.Instance.ForceTaskFactoryOutOfProc && !string.IsNullOrEmpty(_assemblyPath) && FileSystems.Default.FileExists(_assemblyPath))
                 {
                     File.Delete(_assemblyPath);
+                    _assemblyPath = null;
                 }
             }
         }
